refactor: extract timer display logic into reusable TaskTimer component
- Create TaskTimer component to eliminate duplicate timer code
- Remove idle time tracking from App.tsx and FloatingTimer.tsx
- Replace all timer displays with TaskTimer component
- Add appropriate styles for task stack timer display
- Simplify imports by removing unused timer utilities
